cleaned up unused racers and set up ball kart prefab
