--- Assets/PlayerInputAction.inputactions.meta
@@ -1,14 +1,0 @@
-fileFormatVersion: 2
-guid: 1c2c80f274d5e0c4c8f3f957ad5b0785
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
-  generateWrapperCode: 0
-  wrapperCodePath: 
-  wrapperClassName: 
-  wrapperCodeNamespace:
