+++ Assets/Prefabs/BallKart.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 9fc0d4010bbf28b4594072e72b8655ab
-DefaultImporter:
+guid: 49bef55bcb6e097409e1b266925f2eed
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
