--- Assets/Prefabs/Karts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: be7039a223ff70447b837966d4101663
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
