--- Assets/Prefabs/Karts/PhysicsKart.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a00cb00f1a276fe41bd99f1d6d84ea40
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
