--- Assets/Prefabs/Player.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 946eada1524674ba189ff0005ad344d2
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
