--- Assets/Scripts/Control.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 550a7e6849f064f4a932e489d68742d0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
