--- Assets/Scripts/Karts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 406bff1d012d42a4b9ffaf1137391d4d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
