--- Assets/Scripts/Karts/PhysicsKart.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3e18627e92bcc424a975bf27641e014d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
