--- Assets/Scripts/Karts/PhysicsKart/PhysicsCarController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4506a68b896b30945a6d7ba2ad6d3aa2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
